+++ app/src/main/java/com/me/mseotsanyana/mande/BLL/repository/session/iMenuRepository.java
@@ -5,6 +5,6 @@ import com.me.mseotsanyana.mande.DAL.model.session.cMenuModel;
 import java.util.Set;
 
 public interface iMenuRepository {
-    Set<cMenuModel> getMenuModelSet(int userID, int primaryRole,
+    Set<cMenuModel> getMenuModelSet(long userID, int primaryRole,
                                     int secondaryRoles, int statusBITS);
 }
