+++ app/src/main/java/com/me/mseotsanyana/mande/DAL/model/common/cPeriodModel.java
@@ -3,11 +3,11 @@ package com.me.mseotsanyana.mande.DAL.model.common;
 import java.util.Date;
 
 public class cPeriodModel{
-    private int periodID;
-    private int fiscalYearID;
-    private int serverID;
-    private int ownerID;
-    private int orgID;
+    private long periodID;
+    private long fiscalYearID;
+    private long serverID;
+    private long ownerID;
+    private long orgID;
     private int groupBITS;
     private int permsBITS;
     private int statusBITS;
@@ -17,43 +17,43 @@ public class cPeriodModel{
     private Date modifiedDate;
     private Date syncedDate;
 
-    public int getPeriodID() {
+    public long getPeriodID() {
         return periodID;
     }
 
-    public void setPeriodID(int periodID) {
+    public void setPeriodID(long periodID) {
         this.periodID = periodID;
     }
 
-    public int getFiscalYearID() {
+    public long getFiscalYearID() {
         return fiscalYearID;
     }
 
-    public void setFiscalYearID(int fiscalYearID) {
+    public void setFiscalYearID(long fiscalYearID) {
         this.fiscalYearID = fiscalYearID;
     }
 
-    public int getServerID() {
+    public long getServerID() {
         return serverID;
     }
 
-    public void setServerID(int serverID) {
+    public void setServerID(long serverID) {
         this.serverID = serverID;
     }
 
-    public int getOwnerID() {
+    public long getOwnerID() {
         return ownerID;
     }
 
-    public void setOwnerID(int ownerID) {
+    public void setOwnerID(long ownerID) {
         this.ownerID = ownerID;
     }
 
-    public int getOrgID() {
+    public long getOrgID() {
         return orgID;
     }
 
-    public void setOrgID(int orgID) {
+    public void setOrgID(long orgID) {
         this.orgID = orgID;
     }
